update web error view
