+++ app/src/main/java/org/nervos/neuron/activity/AppWebActivity.java
@@ -43,6 +43,7 @@ import org.nervos.neuron.view.WebErrorView;
 import org.nervos.neuron.view.WebMenuPopupWindow;
 import org.nervos.neuron.view.dialog.SignDialog;
 import org.nervos.neuron.view.webview.NeuronWebView;
+import org.nervos.neuron.view.webview.SimpleWebViewClient;
 import org.nervos.neuron.view.webview.item.Address;
 import org.nervos.neuron.view.webview.item.Message;
 import org.nervos.neuron.view.webview.item.Transaction;
@@ -151,7 +152,7 @@ public class AppWebActivity extends NBaseActivity {
                 initViewWhenWebFinish();
             }
         });
-        webView.setWebViewClient(new WebViewClient() {
+        webView.setWebViewClient(new SimpleWebViewClient(webErrorView) {
             @Override
             public boolean shouldOverrideUrlLoading(WebView view, String url) {
                 titleItem = null;
@@ -170,50 +171,6 @@ public class AppWebActivity extends NBaseActivity {
                 }
                 return false;
             }
-
-            @RequiresApi(api = Build.VERSION_CODES.M)
-            @Override
-            public void onReceivedError(WebView view, WebResourceRequest request, WebResourceError error) {
-                try {
-                    switch (error.getErrorCode()) {
-                        case ERROR_AUTHENTICATION:
-                        case ERROR_BAD_URL:
-                        case ERROR_CONNECT:
-                        case ERROR_FAILED_SSL_HANDSHAKE:
-                        case ERROR_HOST_LOOKUP:
-                        case ERROR_PROXY_AUTHENTICATION:
-                        case ERROR_TIMEOUT:
-                        case ERROR_UNKNOWN:
-                            webErrorView.setVisibility(View.VISIBLE);
-                            webView.setVisibility(View.GONE);
-                            break;
-                    }
-                } catch (Exception e) {
-
-                }
-            }
-
-            @Override
-            public void onReceivedError(WebView view, int errorCode, String description, String failingUrl) {
-                try {
-                    switch (errorCode) {
-                        case ERROR_AUTHENTICATION:
-                        case ERROR_BAD_URL:
-                        case ERROR_CONNECT:
-                        case ERROR_FAILED_SSL_HANDSHAKE:
-                        case ERROR_HOST_LOOKUP:
-                        case ERROR_PROXY_AUTHENTICATION:
-                        case ERROR_TIMEOUT:
-                        case ERROR_UNKNOWN:
-                            webErrorView.setVisibility(View.VISIBLE);
-                            webView.setVisibility(View.GONE);
-                            break;
-                    }
-                } catch (Exception e) {
-
-                }
-            }
-
         });
     }
