+++ app/src/main/java/org/nervos/neuron/fragment/AppFragment.java
@@ -29,6 +29,7 @@ import org.nervos.neuron.plugin.AppTabPlugin;
 import org.nervos.neuron.service.httpservice.HttpUrls;
 import org.nervos.neuron.util.web.WebAppUtil;
 import org.nervos.neuron.view.WebErrorView;
+import org.nervos.neuron.view.webview.SimpleWebViewClient;
 
 public class AppFragment extends Fragment {
 
@@ -68,55 +69,13 @@ public class AppFragment extends Fragment {
 
     private void initWebView() {
 
-        webView.setWebViewClient(new WebViewClient() {
+        webView.setWebViewClient(new SimpleWebViewClient(webErrorView) {
             @Override
             public boolean shouldOverrideUrlLoading(WebView view, String url) {
                 Intent intent = new Intent(getContext(), AppWebActivity.class);
                 intent.putExtra(AppWebActivity.EXTRA_URL, url);
                 startActivity(intent);
                 return true;
-            }
-
-            @RequiresApi(api = Build.VERSION_CODES.M)
-            public void onReceivedError(WebView view, WebResourceRequest request, WebResourceError error) {
-                try {
-                    switch (error.getErrorCode()) {
-                        case ERROR_AUTHENTICATION:
-                        case ERROR_BAD_URL:
-                        case ERROR_CONNECT:
-                        case ERROR_FAILED_SSL_HANDSHAKE:
-                        case ERROR_HOST_LOOKUP:
-                        case ERROR_PROXY_AUTHENTICATION:
-                        case ERROR_TIMEOUT:
-                        case ERROR_UNKNOWN:
-                            webErrorView.setVisibility(View.VISIBLE);
-                            webView.setVisibility(View.GONE);
-                            break;
-                    }
-                } catch (Exception e) {
-
-                }
-            }
-
-            @Override
-            public void onReceivedError(WebView view, int errorCode, String description, String failingUrl) {
-                try {
-                    switch (errorCode) {
-                        case ERROR_AUTHENTICATION:
-                        case ERROR_BAD_URL:
-                        case ERROR_CONNECT:
-                        case ERROR_FAILED_SSL_HANDSHAKE:
-                        case ERROR_HOST_LOOKUP:
-                        case ERROR_PROXY_AUTHENTICATION:
-                        case ERROR_TIMEOUT:
-                        case ERROR_UNKNOWN:
-                            webErrorView.setVisibility(View.VISIBLE);
-                            webView.setVisibility(View.GONE);
-                            break;
-                    }
-                } catch (Exception e) {
-
-                }
             }
         });
