+++ app/src/main/java/org/nervos/neuron/service/httpservice/HttpUrls.java
@@ -28,8 +28,8 @@ public class HttpUrls {
 
     // node host
     public static final String APPCHAIN_NODE_IP = "http://121.196.200.225:1337";
-    public static final String ETH_NODE_IP = "https://mainnet.infura.io/h3iIzGIN6msu3KeUrdlt";
-//    public static final String ETH_NODE_IP = "https://rinkeby.infura.io/llyrtzQ3YhkdESt2Fzrk";
+//    public static final String ETH_NODE_IP = "https://mainnet.infura.io/h3iIzGIN6msu3KeUrdlt";
+    public static final String ETH_NODE_IP = "https://rinkeby.infura.io/llyrtzQ3YhkdESt2Fzrk";
 
 
     // discover page config information
