+++ app/src/main/java/org/nervos/neuron/view/WebErrorView.java
@@ -21,7 +21,9 @@ public class WebErrorView extends ConstraintLayout {
         super(context, attrs);
         LayoutInflater.from(context).inflate(R.layout.view_web_error, this);
         ImageView iv = findViewById(R.id.iv);
-        LinearLayout.LayoutParams params = new LinearLayout.LayoutParams((int) (ScreenUtils.getScreenWidth(context) * 0.73), (int) (ScreenUtils.getScreenWidth(context) * 0.73 / 1.4723));
+        LinearLayout.LayoutParams params = new LinearLayout.LayoutParams(
+                (int) (ScreenUtils.getScreenWidth(context) * 0.73),
+                (int) (ScreenUtils.getScreenWidth(context) * 0.73 / 1.4723));
         params.topMargin = (int) (ScreenUtils.getScreenHeight(context) * 0.24);
         iv.setLayoutParams(params);
         findViewById(R.id.btn).setOnClickListener((view) -> {
